refactor(chatCore): extrai transforms de mensagens Claude p/ leaf (#3501)
Lote 3 da decomposição do god-file chatCore.ts (#3501): move os 2 transforms de
mensagens Claude do handleChatCore para leaves, byte-idêntico:

- chatCore/claudeMessageTypes.ts     tipos ClaudeMessage / ClaudeContentBlock (eram
  locais no handler; reimporto só ClaudeMessage, ainda usado num cast).
- chatCore/claudeUpstreamMessages.ts  extractSystemMessagesToBody (PURO) +
  normalizeClaudeUpstreamMessages (captura só `log`; usa extractSystemRoleMessages
  já-leaf + splitMisplacedToolResults). Handler mantém binding fino de `log` ->
  call sites byte-idênticos (1867/1909/1951).

extractSystemMessagesToBody já era DEAD-in-handler (0 call sites; superado por
extractSystemRoleMessages) -> exportado e testado no leaf, NÃO reimportado.
chatCore.ts: 4585 -> 4461 (-124). 1 teste novo (8 casos).
